+++ include/aggregators/SummingAggregator.hpp
@@ -5,7 +5,7 @@ class SummingAggregator : public AggregationFunction {
 	private:
 
 	protected:
-	int aggregateFitnesses(std::vector<int> apportionedFitnesses);
+	float aggregateFitnesses(std::vector<float> apportionedFitnesses);
 
 	public:
 };
