+++ include/selections/RandomSelection.hpp
@@ -13,7 +13,7 @@ class RandomSelection : public SelectionStrategy {
         
 	unsigned int getParent(
 		std::vector<Genome*> population,
-		std::vector<int> fitnesses
+		std::vector<float> fitnesses
 	);
 	std::string toString();
 };
