+++ include/systems/WeakIterativeReplacingGA.hpp
@@ -23,7 +23,7 @@ class WeakIterativeReplacingGA : public ReplacingGA {
 
 	std::vector<Genome*> breedMutateSelect(
 		std::vector<Genome*> initialPopulation,
-		std::vector<int> & populationFitnesses,
+		std::vector<float> & populationFitnesses,
 		CrossoverOperation * cross,
 		MutationOperation * mutation,
 		std::vector<ObjectiveFunction*> objectives,
